requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
alembic==1.18.4
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.11.1
asyncpg==0.31.0
attrs==26.1.0
bcrypt==5.0.0
billiard==4.2.4
celery==5.6.3
certifi==2026.2.25
cffi==2.0.0
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==46.0.6
Django==6.0.3
dnspython==2.8.0
ecdsa==0.19.2
elastic-transport==9.2.1
elasticsearch==9.3.0
email-validator==2.3.0
fastapi==0.135.3
fastapi-cli==0.0.24
fastapi-cloud-cli==0.15.1
fastar==0.9.0
frozenlist==1.8.0
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
kombu==5.6.2
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
minio==7.2.20
multidict==6.7.1
packaging==26.0
passlib==1.7.4
pillow==12.2.0
prompt_toolkit==3.0.52
propcache==0.4.1
psycopg2-binary==2.9.11
pyasn1==0.6.3
pycparser==3.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic-extra-types==2.11.1
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.20.0
pyinstrument==5.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-jose==3.5.0
python-multipart==0.0.22
PyYAML==6.0.3
redis==7.4.0
rich==14.3.3
rich-toolkit==0.19.7
rignore==0.7.6
rsa==4.9.1
sentry-sdk==2.57.0
setuptools==82.0.1
shellingham==1.5.4
simplejson==3.20.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.49
sqlparse==0.5.5
starlette==1.0.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
tzlocal==5.3.1
uliweb-alembic==0.6.9
urllib3==2.6.3
uvicorn==0.42.0
uvloop==0.22.1
vine==5.1.0
watchfiles==1.1.1
wcwidth==0.6.0
websockets==16.0
yarl==1.23.0
